requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
bottleneck=1.3.2=py39h7cc1a96_1
brotlipy=0.7.0=py39hb82d6ee_1003
ca-certificates=2021.10.26=haa95532_2
certifi=2021.10.8=py39haa95532_0
cffi=1.15.0=py39h0878f49_0
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=8.0.3=py39hcbf5309_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py39h7bc7c5c_0
cudatoolkit=11.3.1=h59b6b97_2
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py39haa95532_0
filelock=3.4.0=pyhd8ed1ab_0
freetype=2.10.4=hd328e21_0
googledrivedownloader=0.4=pyhd3deb0d_1
huggingface_hub=0.2.1=pyhd8ed1ab_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.8.2=py39hcbf5309_0
importlib_metadata=4.8.2=hd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py39haa95532_1
ipython=7.29.0=py39hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py39haa95532_1
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h2bbff1b_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
llvmlite=0.37.0=py39h23ce68f_1
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.0.1=py39hb82d6ee_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py39haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkit=9.0=py39h80a8885_2
networkx=2.6.3=pyhd3eb1b0_0
nltk=3.6.5=pyhd8ed1ab_0
numba=0.54.1=py39hf11a4ad_0
numexpr=2.7.3=py39hb80d3ca_1
numpy=1.20.3=py39ha4e8547_0
numpy-base=1.20.3=py39hc2deb75_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1l=h2bbff1b_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=py39h6214cd6_0
pandocfilters=1.4.3=py39haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39hd45dc43_0
pip=21.2.4=py39haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pyg=2.0.2=py39_torch_1.10.0_cu113
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39hcbf5309_4
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.9=2_cp39
pytorch=1.10.0=py3.9_cuda11.3_cudnn8_0
pytorch-cluster=1.5.9=py39_torch_1.10.0_cu113
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.10.0_cu113
pytorch-sparse=0.6.12=py39_torch_1.10.0_cu113
pytorch-spline-conv=1.2.1=py39_torch_1.10.0_cu113
pytz=2021.3=pyhd3eb1b0_0
pywin32=228=py39hbaba5e8_1
pyyaml=6.0=py39hb82d6ee_3
pyzmq=22.3.0=py39hd77b12b_2
regex=2021.11.10=py39hb82d6ee_0
requests=2.26.0=pyhd8ed1ab_1
sacremoses=0.0.46=pyhd8ed1ab_0
scikit-learn=1.0.1=py39hf11a4ad_0
scipy=1.7.1=py39hbe87c03_2
sentence-transformers=2.1.0=pyhd8ed1ab_0
sentencepiece=0.1.95=py39h2e07f2f_1
setuptools=58.0.4=py39haa95532_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
tabulate=0.8.9=py39haa95532_0
tbb=2021.4.0=h59b6b97_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h2bbff1b_0
tokenizers=0.10.3=py39h8c9392a_1
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=py39_cu113
tornado=6.1=py39h2bbff1b_0
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.12.5=pyhd8ed1ab_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.7=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.29.30037=h902a5da_5
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py39hcbf5309_3
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h62dcd97_0
yacs=0.1.6=py_0
yaml=0.2.5=he774522_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
